mod_http_oauth2: Rescope oauth client config into users' storage
This produces client_id of the form owner@host/random and prevents
clients from being deleted by registering an account with the same name
and then deleting the account, as well as having the client
automatically be deleted when the owner account is removed.

On one hand, this leaks the bare JID of the creator to users. On the
other hand, it makes it obvious who made the oauth application.

This module is experimental and only for developers, so this can be
changed if a better method comes up.