mod_post_msg: Add sendxmpp-curl
I wrote this script around the same time as this module.  My local copy
has a last modification timestamp of 2011-03-29, so it's been unchanged
since then.

The original rationale was that it seemed overkill to establish an
entire XMPP session just to send a single message. Nowadays I'm quite
happy using `clix` for this kind task.