mod_pastebin: Fix pasting when <body> is not the first tag (thanks thorsten)
Introduced in ded630a87563

stanza:get_child() and :get_child_text() operates on the first child
when no name or namespace is given.

If it so happens that the first child is an <thread/> instead of <body>
then it calculates the line length against that instead of the body.

Apparently in the vast majority of clients the <body> is first, so this
issue was not noticed until now.