mod_audit: Support for adding location (GeoIP) to audit events
This can be more privacy-friendly than logging full IP addresses, and also
more informative to a user - IP addresses don't mean much to the average
person, however if they see activity from outside their expected country, they
can immediately identify suspicious activity.

As with IPs, this field is configurable for deployments that would like to
disable it. Location is also not logged when the geoip library is not
available.