mod_client_management: Include client software version number in listing
Should you ever wish to revoke a client by version number, e.g. for
security reasons affecting certain versions, then it would be good to at
the very least see which version is used.

Also includes the OAuth2 software ID, an optional unique identifier that
should be the same for all installations of a particular software.