mod_firewall: Update user marks to store instantly via map store
The original approach was to keep marks in memory only, and persist them at
shutdown. That saves I/O, at the cost of potentially losing marks on an
unclean shutdown.

This change persists marks instantly, which may have some performance overhead
but should be more "correct".

It also splits the marking/unmarking into an event which may be watched or
even fired by other modules.