mod_http_oauth2: Make allowed locales configurable
Explicit > Implicit

Instead of allowing anything after #, allow only the explicitly
configured locales to be used.

Default to empty list because using these is not supported yet.

This potentially limits the size of the client_id, which is already
quite large.  Nothing prevents clients from registering a whole
client_id per locale, which would not require translation support on
this side.