mod_measure_modules: Report module statuses via OpenMetrics
Someone in the chat asked about a health check endpoint, which reminded
me of mod_http_status, which provides access to module statuses with
full details. After that, this idea came about, which seems natural.

As noted in the README, it could be used to monitor that critical
modules are in fact loaded correctly.

As more modules use the status API, the more useful this module and
mod_http_status becomes.