mod_pubsub_serverinfo: Warm-up opt-in cache
By warming up the cache that contains the opt-in data, the first publication has a better chance of including domain names for remote domains that opt-in.

Without this change, those domains are named only after the _second_ publication, which can take a while. New users are likely thrown off by that.